requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bidict==0.23.1
certifi==2024.2.2
celery
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.38.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.0
decorator==4.4.2
dj-database-url==2.1.0
Django
django-cors-headers==4.3.1
django-filter==23.5
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
honcho==1.1.0
hyperlink==21.0.0
idna==3.6
imageio==2.34.2
inflection==0.5.1
packaging==24.1
psycopg
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis
requests==2.31.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
tqdm==4.66.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
wcwidth==0.2.13
wsproto==1.2.0
yarl==1.9.11
zope.interface==6.4.post2
